Implement caching, fix Save,Delete.

+++ NeverFoundry.DataStore/PagedList.cs
@@ -24,7 +24,7 @@ namespace NeverFoundry.DataStorage
         /// <summary>
         /// The number of records in the paged query result.
         /// </summary>
-        public long Count => _list.Count;
+        public int Count => _list.Count;
 
         /// <summary>
         /// The zero-based index of the first item in the current page, within the whole collection.
